apps/worker/pyproject.toml
[project]
name = "credit-worker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27.0",
  "psycopg[binary]>=3.2.0",
  "python-dotenv>=1.0.1",
  "pymupdf>=1.24.10",
  "pytesseract>=0.3.10",
  "Pillow>=10.4.0"
]

[project.scripts]
credit-worker = "credit_worker.cli:main"

[tool.ruff]
line-length = 100

[build-system]
requires = ["setuptools>=69.0.0", "wheel>=0.43.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
